Knowledge Graph/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.3
neo4j>=5.18.0
streamlit>=1.35.0
python-dotenv>=1.0.1